fix: GdalDataset
