+++ kitiler-core/src/main/kotlin/dev/tronto/kitiler/core/outgoing/adaptor/gdal/GdalDataset.kt
@@ -2,6 +2,12 @@ package dev.tronto.kitiler.core.outgoing.adaptor.gdal
 
 import dev.tronto.kitiler.core.domain.DataType
 import dev.tronto.kitiler.core.outgoing.adaptor.jts.AffineCoordinateTransform
+import dev.tronto.kitiler.core.outgoing.port.CRS
+import dev.tronto.kitiler.core.outgoing.port.CRSFactory
+import io.github.oshai.kotlinlogging.KotlinLogging
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import org.gdal.gdal.Band
 import org.gdal.gdal.Dataset
 import org.gdal.gdal.gdal
@@ -33,8 +39,37 @@ class GdalDataset(val name: String, val dataset: Dataset, private val memFilePat
     }
 
     override fun close() {
-        sampleBand.delete()
-        dataset.delete()
-        memFilePath?.let { gdal.Unlink(it) }
+        scope.launch {
+            kotlin.runCatching {
+                sampleBand.delete()
+            }
+            kotlin.runCatching {
+                dataset.delete()
+            }
+            memFilePath?.let { gdal.Unlink(it) }
+            // ignore errors
+        }
+    }
+
+    fun getCrs(crsFactory: CRSFactory): CRS {
+        // TODO :: GeoTransform 과 GCP 모두 없는 영상의 CRS 가 잘 만들어지는지 확인.
+        val spatialRef = dataset.GetSpatialRef() ?: dataset.GetGCPSpatialRef()
+        return try {
+            crsFactory.create(spatialRef.ExportToWkt())
+        } finally {
+            try {
+                spatialRef?.delete()
+            } catch (e: Exception) {
+                // ignore
+                logger.warn(e) { "error in delete SpatialReference." }
+            }
+        }
+    }
+
+    companion object {
+        private val scope = CoroutineScope(Dispatchers.IO)
+
+        @JvmStatic
+        private val logger = KotlinLogging.logger { }
     }
 }
